refactor(edit profile page): organize and refactor components on Edit Profile Page
organized and refactored the components on Edit Profile Page to improve the overall structure and
readability of the code. I have grouped the components of the page together to improv readability of
the components structure
