+++ src/components/__tests__/ChangePasswordForm.test.jsx
@@ -1,7 +1,7 @@
 import { I18nContext } from "next-i18next";
 import renderer from "react-test-renderer";
 
-import ChangePasswordForm from "../ChangePasswordForm";
+import ChangePasswordForm from "../editProfile/ChangePasswordForm";
 
 it("renders correctly", () => {
     const i18n = {
