+++ src/components/__tests__/CircleImg.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import CircleImg from "@/components/CircleImg";
+import CircleImg from "@/components/editProfile/CircleImg";
 
 it("renders correctly", () => {
     const tree = renderer.create(<CircleImg />).toJSON();
