+++ src/components/__tests__/EditProfileButton.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import EditProfileButton from "@/components/EditProfileButton";
+import EditProfileButton from "@/components/editProfile/EditProfileButton";
 
 it("renders correctly", () => {
     const tree = renderer.create(<EditProfileButton />).toJSON();
