+++ src/components/__tests__/EditProfileForm.test.jsx
@@ -1,7 +1,7 @@
 import { I18nContext } from "next-i18next";
 import renderer from "react-test-renderer";
 
-import EditProfileForm from "../EditProfileForm";
+import EditProfileForm from "../editProfile/EditProfileForm";
 
 it("renders correctly", () => {
     const i18n = {
