+++ src/components/__tests__/SaveButton.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import SaveButton from "@/components/SaveButton";
+import SaveButton from "@/components/editProfile/SaveButton";
 
 it("renders correctly", () => {
     const tree = renderer.create(<SaveButton />).toJSON();
