+++ src/components/__tests__/SecondaryTitle.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import SecondaryTitle from "@/components/SecondaryTitle";
+import SecondaryTitle from "@/components/editProfile/SecondaryTitle";
 
 it("renders correctly", () => {
     const tree = renderer.create(<SecondaryTitle />).toJSON();
