+++ src/components/__tests__/SingleInterest.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import SingleInterest from "../SingleInterest";
+import SingleInterest from "../editProfile/SingleInterest";
 
 it("renders correctly", () => {
     const tree = renderer.create(<SingleInterest />).toJSON();
